Ran all sourse through JEDI code formater.
[mainFrom.pas]
Added a function to display the filesize in a Human readable format.
Cleaned up the code in several places where nested it .. then exits.

[formAbout.pas]
Amneded, so that is knows if it was compiled in either 32
or 64 bit mode.
Added a powered by lazarus badge.

[uInfor.pas]
Added constants used by the FileSizeToHumanReadableString finction in
mainForm.pas.